requirements.txt
transformers
huggingface_hub
streamlit
python-dotenv